Fix the MSVC build: strncpy is deprecated there, and /WX means fatal
island.c's SDL_strlcpy replacement used strncpy, which MSVC deprecates
(C4996) — so the Windows job failed at the sim library while Linux and
macOS were green. building.c's set_reason() had already learned this
exact lesson; snprintf is the project's answer, and it null-terminates
on truncation, which strncpy does not.

Two more hazards in code the Windows job never reached, since it died
before compiling net.c or the server:

  - net.c called Sleep() but no longer includes SDL, which is what used
    to drag in windows.h. Now included explicitly, after winsock2.h.
  - the server printed a uint16_t port with %u; it promotes to int, and
    MSVC checks printf format types at /W4 (C4477). Cast.

+++ server/saltmarch_host.c
@@ -167,7 +167,8 @@ int main(int argc, char *argv[])
 
     ns = net_host(port);
     if (!ns) {
-        fprintf(stderr, "host: could not listen on port %u\n", port);
+        fprintf(stderr, "host: could not listen on port %u\n",
+                (unsigned)port);
         game_free(gs);
         return 1;
     }
@@ -178,7 +179,7 @@ int main(int argc, char *argv[])
 
     start_tick = gs->sim_tick_no;
     printf("host: listening on %u, ticking at %d ms, checkpointing to %s\n",
-           port, SIM_TICK_MS, world_path);
+           (unsigned)port, SIM_TICK_MS, world_path);
     fflush(stdout);
 
     last_ns = now_ns();
